requirements.txt
Flask
requests
python-dotenv
speechrecognition
pyttsx3
